o  fixed refereces to ints.
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497958 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-core/src/main/java/org/apache/maven/continuum/core/action/ExecuteBuilderContinuumAction.java
@@ -127,7 +127,7 @@ public class ExecuteBuilderContinuumAction
 
         store.addBuildResult( project, buildResult );
 
-        context.put( KEY_BUILD_ID, Integer.toString( buildResult.getId() ) );
+        context.put( KEY_BUILD_ID, Long.toString( buildResult.getId() ) );
 
         buildResult = store.getBuildResult( buildResult.getId() );
